+++ src/pages/HomePage/types/index.ts
@@ -1,19 +1,3 @@
-// theme 카테고리 섹션
-export type ThemeData = {
-  id: number;
-  key: string;
-  label: string;
-  imageURL?: string;
-  title: string;
-  description?: string;
-  backgroundColor?: string;
-};
-
-export type GetThemesResponse = {
-  themes: ThemeData[];
-};
-
-// 실시간 급상승 선물랭킹 섹션
 export type RankingFilter = {
   targetType: 'ALL' | 'FEMALE' | 'MALE' | 'TEEN';
   rankType: 'MANY_WISH' | 'MANY_RECEIVE' | 'MANY_WISH_RECEIVE';
