+++ src/types/themeType.ts
@@ -1,18 +1,24 @@
-export type ThemeHeaderType = {
-  id: number;
-  themeTitle: string;
-  subTitle: string;
-  description: string;
-  color: string;
-};
-
 export interface ThemeListType {
   id: number;
   subtitle: string;
   title: string;
-  amount: string;
+  amount: number;
 }
 
 export interface ThemeListDataType {
   [key: string]: ThemeListType[];
 }
+
+export type ThemeData = {
+  id: number;
+  key: string;
+  label: string;
+  imageURL?: string;
+  title: string;
+  description?: string;
+  backgroundColor?: string;
+};
+
+export type GetThemesResponse = {
+  themes: ThemeData[];
+};
